refactor: 💡 WIP: Started Mongo refactor
