+++ index.js
@@ -3,16 +3,12 @@
 require('dotenv').config();
 //const socketServer = require('./src/socketServer');
 const server = require('./src/server.js');
-// mongoose 
-const mongoose = require('mongoose');
-mongoose.connect(process.env.MONGOOSE_URI, { useNewUrlParser: true, useUnifiedTopology: true });
+const mongo = require('./src/database/databaseInit.js')
+
+
 
 console.log('\nStarting Servers and Database:');
-
-const db = mongoose.connection;
-db.on('error', console.error.bind(console, 'conection error: '));
-db.once('open', () => console.log('(3/3) MongoDB Connected'));
-
+mongo.start();
 // socketServer.start(process.env.SOCKETPORT);
 server.start(process.env.PORT, process.env.SOCKETPORT);
